Introduce load_tests_for_base_classes
