+++ Lib/test/test_decimal/test_other.py
@@ -5018,21 +5018,5 @@ class SignatureTest(unittest.TestCase):
         doit('Context')
 
 
-class TestModule:
-    def test_deprecated__version__(self):
-        with self.assertWarnsRegex(
-            DeprecationWarning,
-            "'__version__' is deprecated and slated for removal in Python 3.20",
-        ) as cm:
-            getattr(self.decimal, "__version__")
-        self.assertEqual(cm.filename, __file__)
-
-
-@requires_cdecimal
-class CTestModule(TestModule, unittest.TestCase):
-    decimal = C
-class PyTestModule(TestModule, unittest.TestCase):
-    decimal = P
-
 if __name__ == '__main__':
     unittest.main()
